requirements.txt
attrs~=21.2.0
marshmallow~=3.13.0
requests~=2.26.0
lxml~=4.6.3
openpyxl~=3.0.10